Add variant script
This scripts modifies the schematic according to the match of a field
with a given value. The matching parts are changed to their "_NOPOP"
counterparts. The resulting schematic then shows the not populated parts
with some special symbol.

BEWARE ! This change is destructive! there is yet no inverting
commmand. This command should only be used if you have a backup of your
schematic files or a version committed in your revision control system.