Make variant.py switchable
if you specify a different field for variant, the schematic is switched
to this new field. Incidentally, specifying a non-existent field should
switch the schematic back to the original setup.

Beware that the file is fully filtered, hence it is modified even if the
graphics of the schematic seems identical.